refactor: split dynamic compatibility semantics
