+++ contractAcquisition/src/main/kotlin/dev/obiente/nextcloudnative/contracts/StaticRouteContract.kt
@@ -2,7 +2,6 @@ package dev.obiente.nextcloudnative.contracts
 
 import org.json.JSONArray
 import org.json.JSONObject
-import java.security.MessageDigest
 
 /**
  * Builds a deliberately conservative contract from static Nextcloud route metadata.
@@ -627,7 +626,7 @@ private val SEMANTIC_PATH_NOISE = setOf(
     "api", "apps", "indexphp", "ocs", "ocsv1php", "ocsv2php", "list", "index", "getall", "findall",
 )
 
-private data class StaticRoute(
+internal data class StaticRoute(
     val controller: String,
     val method: String,
     val url: String,
@@ -636,7 +635,7 @@ private data class StaticRoute(
     val requirements: Map<String, String>,
 )
 
-private data class StaticApiController(
+internal data class StaticApiController(
     val name: String,
     val methods: Set<String>,
     val singleParameters: Map<String, StaticPhpParameter>,
@@ -647,7 +646,7 @@ private data class StaticApiController(
     val normalizedName: String = name.removeSuffix("Controller").normalizedPhpName()
 }
 
-private data class StaticPhpParameter(
+internal data class StaticPhpParameter(
     val name: String,
     val type: String,
     val required: Boolean = true,
@@ -661,13 +660,6 @@ private data class StaticSettingsSetter(
 
 private data class StaticCrudWrite(
     val bodyParameters: List<StaticPhpParameter>,
-)
-
-private data class VerifiedExactWrite(
-    val label: String,
-    val resourceId: String,
-    val bodySchema: JSONObject?,
-    val required: Boolean = true,
 )
 
 private fun isVerifiedChoresController(
@@ -898,143 +890,7 @@ private fun verifiedChoresWrite(
     }
 }
 
-private fun isVerifiedBudgetAccountController(
-    appId: String,
-    appVersion: String,
-    controller: StaticApiController,
-    controllerSource: String?,
-): Boolean =
-    appId == "budget" && appVersion == "2.44.0" &&
-        controller.normalizedName == "account" && controllerSource != null &&
-        controllerSource.sha256() == BUDGET_2_44_0_ACCOUNT_CONTROLLER_SHA256
-
-/**
- * Imports the deliberately small account mutation surface proven by Budget 2.44.0's signed
- * AccountController. The controller reads IRequest directly, so its safe body cannot be derived
- * from the PHP signature. The digest makes the exception fail closed when that implementation
- * changes; optional banking and liability fields remain out of the generic editor.
- */
-private fun verifiedBudgetAccountWrite(
-    appId: String,
-    appVersion: String,
-    route: StaticRoute,
-    fullPath: String,
-    controller: StaticApiController,
-    controllerSource: String?,
-): VerifiedExactWrite? {
-    if (!isVerifiedBudgetAccountController(appId, appVersion, controller, controllerSource)) {
-        return null
-    }
-    val editableProperties = mapOf(
-        "name" to stringSchema(title = "Account name"),
-        "type" to stringSchema(title = "Account type").put(
-            "enum",
-            JSONArray(
-                listOf(
-                    "checking",
-                    "savings",
-                    "credit_card",
-                    "investment",
-                    "loan",
-                    "cash",
-                    "money_market",
-                    "cryptocurrency",
-                    "mortgage",
-                    "line_of_credit",
-                ),
-            ),
-        ),
-        "currency" to stringSchema(title = "Currency code"),
-        "institution" to stringSchema(title = "Institution"),
-    )
-    return when (Triple(route.method.normalizedPhpName(), route.verb, fullPath)) {
-        Triple("create", "POST", "/apps/budget/api/accounts") -> VerifiedExactWrite(
-            label = "Create account",
-            resourceId = "accounts",
-            bodySchema = closedObjectSchema(
-                properties = editableProperties +
-                    ("balance" to JSONObject().put("type", "number").put("title", "Opening balance")),
-                required = listOf("name", "type"),
-            ),
-        )
-        Triple("update", "PUT", "/apps/budget/api/accounts/{id}") -> VerifiedExactWrite(
-            label = "Edit account",
-            resourceId = "accounts",
-            bodySchema = closedObjectSchema(
-                properties = editableProperties,
-                required = emptyList(),
-            ),
-            required = false,
-        )
-        Triple("destroy", "DELETE", "/apps/budget/api/accounts/{id}") -> VerifiedExactWrite(
-            label = "Delete account",
-            resourceId = "accounts",
-            bodySchema = null,
-            required = false,
-        )
-        else -> null
-    }
-}
-
-private fun isVerifiedMusicPlaylistController(
-    appId: String,
-    appVersion: String,
-    controller: StaticApiController,
-    controllerSource: String?,
-): Boolean =
-    appId == "music" && appVersion == "3.1.1" &&
-        controller.normalizedName == "playlistapi" && controllerSource != null &&
-        controllerSource.sha256() == MUSIC_3_1_1_PLAYLIST_API_CONTROLLER_SHA256
-
-/**
- * Imports the scalar playlist lifecycle verified against Music 3.1.1. The controller's required
- * nullable union for trackIds is not representable by the general PHP signature parser, although
- * the implementation and live API both accept it as null. New playlists therefore expose only a
- * required name; track membership stays on the app's dedicated, identity-bound actions.
- */
-private fun verifiedMusicPlaylistWrite(
-    appId: String,
-    appVersion: String,
-    route: StaticRoute,
-    fullPath: String,
-    controller: StaticApiController,
-    controllerSource: String?,
-): VerifiedExactWrite? {
-    if (!isVerifiedMusicPlaylistController(appId, appVersion, controller, controllerSource)) {
-        return null
-    }
-    return when (Triple(route.method.normalizedPhpName(), route.verb, fullPath)) {
-        Triple("create", "POST", "/apps/music/api/playlists") -> VerifiedExactWrite(
-            label = "Create playlist",
-            resourceId = "playlists",
-            bodySchema = closedObjectSchema(
-                properties = mapOf("name" to stringSchema(title = "Playlist name")),
-                required = listOf("name"),
-            ),
-        )
-        Triple("update", "PUT", "/apps/music/api/playlists/{id}") -> VerifiedExactWrite(
-            label = "Edit playlist",
-            resourceId = "playlists",
-            bodySchema = closedObjectSchema(
-                properties = mapOf(
-                    "name" to stringSchema(title = "Playlist name"),
-                    "comment" to stringSchema(title = "Comment"),
-                ),
-                required = emptyList(),
-            ),
-            required = false,
-        )
-        Triple("delete", "DELETE", "/apps/music/api/playlists/{id}") -> VerifiedExactWrite(
-            label = "Delete playlist",
-            resourceId = "playlists",
-            bodySchema = null,
-            required = false,
-        )
-        else -> null
-    }
-}
-
-private fun closedObjectSchema(
+internal fun closedObjectSchema(
     properties: Map<String, JSONObject>,
     required: List<String>,
 ): JSONObject = JSONObject()
@@ -1044,11 +900,6 @@ private fun closedObjectSchema(
     .also { schema ->
         if (required.isNotEmpty()) schema.put("required", JSONArray(required))
     }
-
-private fun stringSchema(title: String, format: String? = null): JSONObject = JSONObject()
-    .put("type", "string")
-    .put("title", title)
-    .also { schema -> format?.let { schema.put("format", it) } }
 
 private fun integerSchema(title: String): JSONObject = JSONObject()
     .put("type", "integer")
@@ -1084,16 +935,8 @@ private fun choresRepeatScheduleSchema(): JSONObject {
         .put("x-nextcloud-native-enum-labels", JSONObject(choices))
 }
 
-private fun String.sha256(): String = MessageDigest.getInstance("SHA-256")
-    .digest(encodeToByteArray())
-    .joinToString("") { byte -> (byte.toInt() and 0xff).toString(16).padStart(2, '0') }
-
 private const val CHORES_0_1_0_API_CONTROLLER_SHA256 =
     "146286dcb68bddd025e0a47e7edc134fbc94f0e9f594e9030663bb0f217f3cc6"
-private const val BUDGET_2_44_0_ACCOUNT_CONTROLLER_SHA256 =
-    "f873d3d8dc640c8baa05d85c34e4a08e0eb62846c5dd9ca7f8e07e495d0caf7d"
-private const val MUSIC_3_1_1_PLAYLIST_API_CONTROLLER_SHA256 =
-    "f8146b8521487e79a8ae2ba4ce6eec5556d3bf4e596d00dfcfbe2b5d5bd5877b"
 
 /**
  * Proves only conventional scalar CRUD signatures. Every route placeholder and every required
@@ -1805,7 +1648,7 @@ private fun String.isSafeStaticJsonReturnType(): Boolean {
     return !endsWith("Response") && this !in STATIC_NON_DATA_RETURN_TYPES
 }
 
-private fun String.normalizedPhpName(): String = lowercase().filter(Char::isLetterOrDigit)
+internal fun String.normalizedPhpName(): String = lowercase().filter(Char::isLetterOrDigit)
 
 private fun String.humanizedPhpName(): String = buildString(length + 4) {
     this@humanizedPhpName.forEachIndexed { index, character ->
