+++ ui/src/commonTest/kotlin/dev/obiente/nextcloudnative/nativeui/model/DynamicNavigationPlannerTest.kt
@@ -13,31 +13,6 @@ import kotlin.test.assertTrue
 
 class DynamicNavigationPlannerTest {
     @Test
-    fun `GET helper requiring a body key cannot become an app root`() {
-        val helper = action(
-            id = "photo-by-key",
-            resourceId = "contacts",
-            intent = ActionIntent.read,
-            body = HttpBody(
-                contentType = "application/json",
-                required = true,
-                schema = buildJsonObject {
-                    put("required", buildJsonArray { add(JsonPrimitive("key")) })
-                },
-            ),
-        )
-        val descriptor = hierarchyDescriptor().copy(
-            resources = listOf(resource("contacts")),
-            layouts = listOf(layout("contacts", helper.id)),
-            actions = listOf(helper),
-            links = emptyList(),
-            forms = emptyList(),
-        )
-
-        assertTrue(descriptor.planDynamicNavigation().rootDestinations.isEmpty())
-    }
-
-    @Test
     fun `interactive autocomplete collection is a picker source rather than an app root`() {
         val autocomplete = action(
             id = "user-autocomplete",
@@ -123,110 +98,6 @@ class DynamicNavigationPlannerTest {
         )
         assertFalse(plan.contextualFormActions.any { it.formId == "edit-bill.form" })
         assertFalse(plan.contextualFormActions.any { it.formId == "create-project.form" })
-    }
-
-    @Test
-    fun `verified self contained app command is available without a matching read root`() {
-        val listProjects = action("list-projects", "projects", ActionIntent.list)
-        val importData = action(
-            "import-data",
-            "imports",
-            ActionIntent.execute,
-            method = HttpMethod.POST,
-        ).copy(
-            confidence = Confidence.high,
-            provenance = listOf(
-                Provenance(
-                    kind = ProvenanceKind.appStoreLinkedSourceTag,
-                    source = "signed catalog release",
-                    detail = "Exact source tag contract",
-                ),
-            ),
-        )
-        val importForm = form(
-            "import-data.form",
-            "Import data",
-            "imports",
-            importData.id,
-        ).copy(
-            confidence = Confidence.high,
-            provenance = importData.provenance,
-        )
-        val descriptor = hierarchyDescriptor().copy(
-            resources = listOf(resource("projects"), resource("imports")),
-            layouts = listOf(layout("projects", listProjects.id)),
-            links = emptyList(),
-            forms = listOf(importForm),
-            actions = listOf(listProjects, importData),
-        )
-
-        assertEquals(
-            listOf(importForm.id),
-            descriptor.planDynamicNavigation().rootFormActions.map(DynamicNavigationFormAction::formId),
-        )
-        assertTrue(
-            descriptor.copy(
-                forms = listOf(importForm.copy(provenance = emptyList())),
-            ).planDynamicNavigation().rootFormActions.isEmpty(),
-        )
-        assertTrue(
-            descriptor.copy(
-                actions = listOf(
-                    listProjects,
-                    importData.copy(
-                        binding = importData.binding.copy(
-                            path = "/imports/{projectId}",
-                            pathParameters = listOf(
-                                HttpParameter(
-                                    "projectId",
-                                    required = true,
-                                    schema = buildJsonObject {},
-                                    source = ParameterSource.resourceField,
-                                ),
-                            ),
-                        ),
-                    ),
-                ),
-            ).planDynamicNavigation().rootFormActions.isEmpty(),
-        )
-    }
-
-    @Test
-    fun `root forms with the same label remain distinct across resources`() {
-        val evidence = listOf(
-            Provenance(
-                kind = ProvenanceKind.verifiedAppPackage,
-                source = "signed app package",
-                detail = "Verified root writes",
-            ),
-        )
-        val createAccount = action(
-            "create-account",
-            "accounts",
-            ActionIntent.create,
-            method = HttpMethod.POST,
-        ).copy(label = "Create", provenance = evidence)
-        val createCategory = action(
-            "create-category",
-            "categories",
-            ActionIntent.create,
-            method = HttpMethod.POST,
-        ).copy(label = "Create", provenance = evidence)
-        val descriptor = hierarchyDescriptor().copy(
-            resources = listOf(resource("accounts"), resource("categories")),
-            layouts = emptyList(),
-            links = emptyList(),
-            forms = listOf(
-                form("create-account.form", "Create", "accounts", createAccount.id).copy(provenance = evidence),
-                form("create-category.form", "Create", "categories", createCategory.id).copy(provenance = evidence),
-            ),
-            actions = listOf(createAccount, createCategory),
-        )
-
-        assertEquals(
-            listOf("create-account.form", "create-category.form"),
-            descriptor.planDynamicNavigation().rootFormActions.map(DynamicNavigationFormAction::formId),
-        )
     }
 
     @Test
