+++ ui/src/commonTest/kotlin/dev/obiente/nextcloudnative/nativeui/runtime/GenericNativeRendererStateTest.kt
@@ -2261,87 +2261,6 @@ class GenericNativeRendererStateTest {
     }
 
     @Test
-    fun `required read only body identity may reuse the exact required path identity`() {
-        val bodySchema = Json.parseToJsonElement(
-            """{
-              "type":"object",
-              "properties":{
-                "id":{"type":"string","readOnly":true},
-                "name":{"type":"string"}
-              },
-              "required":["id","name"],
-              "additionalProperties":false
-            }""",
-        )
-        val resource = ResourceSpec(
-            id = "recipes",
-            name = "Recipes",
-            confidence = Confidence.high,
-            fields = listOf(
-                field(id = "id", kind = FieldKind.string, required = true),
-                field(id = "name", kind = FieldKind.string, required = true),
-            ),
-        )
-        val update = action(
-            id = "recipes.update",
-            intent = ActionIntent.update,
-            risk = ActionRisk.mutating,
-            method = HttpMethod.PUT,
-        ).copy(
-            resourceId = resource.id,
-            binding = ApiBinding(
-                method = HttpMethod.PUT,
-                path = "/recipes/{id}",
-                operationId = "recipes.update",
-                pathParameterNames = listOf("id"),
-                requiredPathParameterNames = listOf("id"),
-                bodyFieldNames = listOf("id", "name"),
-                requiredBodyFieldNames = listOf("id", "name"),
-                bodyContentType = "application/json",
-                bodySchema = bodySchema,
-            ),
-            inputSchema = bodySchema,
-        )
-        val view = ViewSpec(
-            id = "recipes.update.form",
-            title = "Update recipe",
-            resourceId = resource.id,
-            component = NativeComponent.form,
-            sourceActionId = update.id,
-            confidence = Confidence.high,
-        )
-        val schema = NativeAppSchema(
-            schemaVersion = "0.1",
-            app = AppIdentity("cookbook", "Cookbook", "0.11.10"),
-            confidence = Confidence.high,
-            resources = listOf(resource),
-            views = listOf(view),
-            actions = listOf(update),
-        )
-
-        val ready = assertIs<NativeRequestBuildResult.Ready>(
-            buildNativeSubmitRequest(
-                schema = schema,
-                view = view,
-                values = mapOf("id" to "208", "name" to "Updated recipe"),
-                confirmed = false,
-            ),
-        )
-        assertEquals(
-            mapOf("id" to "208", "name" to "Updated recipe"),
-            assertIs<NativeActionRequest.Submit>(ready.request).values,
-        )
-        assertIs<NativeRequestBuildResult.Invalid>(
-            buildNativeSubmitRequest(
-                schema = schema,
-                view = view,
-                values = mapOf("name" to "Updated recipe"),
-                confirmed = false,
-            ),
-        )
-    }
-
-    @Test
     fun `generic create submit binds one structurally proven normalized parent id`() {
         val resource = ResourceSpec(
             id = "lists",
